Make p4.el play nice(r) with TRAMP
Replace use of call-process and start-process with the tramp-aware variants
process-file and start-file-process

Use file-local-name transformation of buffer-file-name in p4-buffer-file-name,
and use p4-buffer-file-name instead of buffer-file-name in
p4-update-status. This might not be correct.

Added a comment to remind myself to implement a tramp-aware call-process-region

This still falls over with at least some of the synchronous operations, but
assync seems to work fairly well.